COMP: Fix sign-compare warnings

+++ libsrc/ImageSEGConverter.cpp
@@ -112,7 +112,7 @@ namespace dcmqi {
     int uidfound = 0, uidnotfound = 0;
 
     Uint8 *frameData = new Uint8[frameSize];
-    for(int segFileNumber=0; segFileNumber<segmentations.size(); segFileNumber++){
+    for(size_t segFileNumber=0; segFileNumber<segmentations.size(); segFileNumber++){
 
       cout << "Processing input label " << segmentations[segFileNumber] << endl;
 
@@ -157,7 +157,7 @@ namespace dcmqi {
         return NULL;
       }
 
-      for(int segLabelNumber=0 ; segLabelNumber<l2lm->GetOutput()->GetNumberOfLabelObjects();segLabelNumber++){
+      for(unsigned segLabelNumber=0 ; segLabelNumber<l2lm->GetOutput()->GetNumberOfLabelObjects();segLabelNumber++){
         LabelType* labelObject = l2lm->GetOutput()->GetNthLabelObject(segLabelNumber);
         short label = labelObject->GetLabel();
 
@@ -252,7 +252,7 @@ namespace dcmqi {
 
         // TODO: make it possible to skip empty frames (optional)
         // iterate over slices for an individual label and populate output frames
-        for(int sliceNumber=firstSlice;sliceNumber<lastSlice;sliceNumber++){
+        for(unsigned sliceNumber=firstSlice;sliceNumber<lastSlice;sliceNumber++){
 
           // PerFrame FG: FrameContentSequence
           //fracon->setStackID("1"); // all frames go into the same stack
@@ -331,7 +331,9 @@ namespace dcmqi {
             }
 
             OFVector<DcmDataset*> siVector;
-            for(unsigned derImageInstanceNum=0;derImageInstanceNum<slice2derimg[sliceNumber].size();derImageInstanceNum++){
+            for(size_t derImageInstanceNum=0;
+                derImageInstanceNum<slice2derimg[sliceNumber].size();
+                derImageInstanceNum++){
               siVector.push_back(dcmDatasets[slice2derimg[sliceNumber][derImageInstanceNum]]);
             }
 
@@ -500,7 +502,7 @@ namespace dcmqi {
 
     populateMetaInformationFromDICOM(segDataset, segdoc, metaInfo);
 
-    for(int frameId=0;frameId<fgInterface.getNumberOfFrames();frameId++){
+    for(size_t frameId=0;frameId<fgInterface.getNumberOfFrames();frameId++){
       const DcmIODTypes::Frame *frame = segdoc->getFrame(frameId);
       bool isPerFrame;
 
@@ -646,8 +648,8 @@ namespace dcmqi {
         unpackedFrame = new DcmIODTypes::Frame(*frame);
 
       // initialize slice with the frame content
-      for(int row=0;row<imageSize[1];row++){
-        for(int col=0;col<imageSize[0];col++){
+      for(unsigned row=0;row<imageSize[1];row++){
+        for(unsigned col=0;col<imageSize[0];col++){
           ImageType::PixelType pixel;
           unsigned bitCnt = row*imageSize[0]+col;
           pixel = unpackedFrame->pixData[bitCnt];
@@ -675,7 +677,7 @@ namespace dcmqi {
     // Assume that orientation of the segmentation is the same as the source series
     unsigned numLabelSlices = labelImage->GetLargestPossibleRegion().GetSize()[2];
     vector<vector<int> > slice2derimg(numLabelSlices);
-    for(int i=0;i<dcmDatasets.size();i++){
+    for(size_t i=0;i<dcmDatasets.size();i++){
       OFString ippStr;
       ImageType::PointType ippPoint;
       ImageType::IndexType ippIndex;
